adding stac catalog...
