+++ sites/geohub/src/lib/config/AppConfig/StacCatalogs.ts
@@ -1,0 +1,12 @@
+export const StacCatalogs = [
+	{
+		id: 'maxar-opendata',
+		name: 'Maxar Open Data',
+		url: `https://maxar-opendata.s3.amazonaws.com/events/catalog.json`
+	},
+	{
+		id: 'nz-imagery',
+		name: 'New Zealand Imagery',
+		url: `https://nz-imagery.s3-ap-southeast-2.amazonaws.com/catalog.json`
+	}
+];
