+++ sites/geohub/src/routes/(app)/management/stac/api/+page.server.ts
