+++ sites/geohub/src/routes/(app)/management/stac/api/+page.ts
@@ -2,12 +2,12 @@ import { StacApis } from '$lib/config/AppConfig';
 import type { PageLoad } from './$types';
 
 export const load: PageLoad = async () => {
-	const title = 'STAC management | GeoHub';
-	const content = 'STAC management';
+	const title = 'STAC API management | GeoHub';
+	const content = 'STAC API management';
 
 	return {
 		title,
 		content,
-		stacs: StacApis
+		stacApis: StacApis
 	};
 };
