+++ sites/geohub/src/routes/(app)/management/stac/api/[type]/[collection]/+page.ts
